+++ rstar/src/algorithm/mod.rs
@@ -1,5 +1,6 @@
 pub mod bulk_load;
 pub mod intersection_iterator;
+/// Iterator types
 pub mod iterators;
 pub mod nearest_neighbor;
 pub mod removal;