+++ rstar/src/primitives/mod.rs
@@ -1,9 +1,11 @@
 //! Contains primitives ready for insertion into an r-tree.
 
+mod geom_with_data;
 mod line;
 mod point_with_data;
 mod rectangle;
 
+pub use self::geom_with_data::GeomWithData;
 pub use self::line::Line;
 pub use self::point_with_data::PointWithData;
 pub use self::rectangle::Rectangle;